# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ASEP Engine",
    "description": "AI engine for the ASEP platform: chat, agents, repository intelligence.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/repositories/{repository_id}/work-items": {
      "post": {
        "summary": "Create Work Item",
        "operationId": "create_work_item_v1_repositories__repository_id__work_items_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkItemIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List Work Items",
        "operationId": "list_work_items_v1_repositories__repository_id__work_items_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemOut"
                  },
                  "title": "Response List Work Items V1 Repositories  Repository Id  Work Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{repository_id}/work-items/{item_id}": {
      "patch": {
        "summary": "Update Work Item",
        "operationId": "update_work_item_v1_repositories__repository_id__work_items__item_id__patch",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{repository_id}/work-items/{item_id}/push": {
      "post": {
        "summary": "Push Work Item",
        "description": "Create an issue in a connected tracker from this work item and store the\nlink. 404 when the tracker is not connected.",
        "operationId": "push_work_item_v1_repositories__repository_id__work_items__item_id__push_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{repository_id}/work-items/reorder": {
      "post": {
        "summary": "Reorder Work Items",
        "description": "Set each item's board position from its index in `ordered_ids`.",
        "operationId": "reorder_work_items_v1_repositories__repository_id__work_items_reorder_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemOut"
                  },
                  "title": "Response Reorder Work Items V1 Repositories  Repository Id  Work Items Reorder Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{repository_id}/roadmap": {
      "post": {
        "summary": "Generate Repository Roadmap",
        "description": "Scrum Master: turn a one-line goal into work items saved to the backlog.",
        "operationId": "generate_repository_roadmap_v1_repositories__repository_id__roadmap_post",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoadmapIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkItemOut"
                  },
                  "title": "Response Generate Repository Roadmap V1 Repositories  Repository Id  Roadmap Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/repositories/{repository_id}/work-items/insights": {
      "get": {
        "summary": "Work Item Insights",
        "description": "Blocked items and the recommended next item \u2014 computed, never stored.",
        "operationId": "work_item_insights_v1_repositories__repository_id__work_items_insights_get",
        "parameters": [
          {
            "name": "repository_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Repository Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanInsightsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Estimate": {
        "type": "string",
        "enum": [
          "small",
          "medium",
          "large"
        ],
        "title": "Estimate",
        "description": "Relative size only \u2014 never a false-precision hour count."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanInsightsOut": {
        "properties": {
          "blocked": {
            "items": {
              "$ref": "#/components/schemas/BlockedItemOut"
            },
            "type": "array",
            "title": "Blocked"
          },
          "recommended": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkItemOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "blocked",
          "recommended"
        ],
        "title": "PlanInsightsOut"
      },
      "Priority": {
        "type": "string",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Priority"
      },
...
      "PushIn": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/IntegrationKind",
            "default": "linear"
          }
        },
        "type": "object",
        "title": "PushIn",
        "description": "Which connected issue tracker to push the work item to."
      },
      "ReorderIn": {
        "properties": {
          "ordered_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ordered Ids"
          }
        },
        "type": "object",
        "required": [
          "ordered_ids"
        ],
        "title": "ReorderIn"
      },
...
      "RoadmapIn": {
        "properties": {
          "goal": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Goal"
          }
        },
        "type": "object",
        "required": [
          "goal"
        ],
        "title": "RoadmapIn"
      },
...
      "WorkItemIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkItemKind",
            "default": "feature"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority",
            "default": "medium"
          },
          "estimate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Estimate"
              },
              {
                "type": "null"
              }
            ]
          },
          "milestone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone"
          },
          "depends_on": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Depends On"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "WorkItemIn"
      },
      "WorkItemKind": {
        "type": "string",
        "enum": [
          "feature",
          "bug",
          "chore",
          "spike"
        ],
        "title": "WorkItemKind"
      },
      "WorkItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "repository_id": {
            "type": "string",
            "format": "uuid",
            "title": "Repository Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "estimate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "milestone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "implemented_by_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implemented By Run Id"
          },
          "external_issue_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Issue Url"
          },
          "external_issue_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Issue Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "repository_id",
          "title",
          "description",
          "kind",
          "status",
          "estimate",
          "priority",
          "milestone",
          "depends_on",
          "rationale",
          "position",
          "implemented_by_run_id",
          "external_issue_url",
          "external_issue_key",
          "created_at",
          "updated_at"
        ],
        "title": "WorkItemOut"
      },
...
      "WorkItemUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "kind": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkItemKind"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkItemStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Priority"
              },
              {
                "type": "null"
              }
            ]
          },
          "estimate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Estimate"
              },
              {
                "type": "null"
              }
            ]
          },
          "milestone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone"
          },
          "depends_on": {
            "anyOf": [
              {
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depends On"
          },
          "rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale"
          },
          "position": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          }
        },
        "type": "object",
        "title": "WorkItemUpdate",
        "description": "Every field optional; only the fields sent are changed (partial update)."
      }
    }
  }
}